requirements.txt
Django==2.0
django-ckeditor==5.4.0
django-comments-xtd==2.0.9
django-contrib-comments==1.8.0
django-js-asset==0.1.1
django-ranged-response==0.2.0
django-registration-redux==2.0
django-render-block==0.5
django-simple-captcha==0.5.6
djangoajax==2.4
djangorestframework==3.7.3
docutils==0.14
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
